Move debian logic to .debian.followup.
